+++ react/components/MiniCartContent.js
@@ -126,8 +126,8 @@ class MiniCartContent extends Component {
   }
 
   renderWithoutItems = label => (
-    <div className="vtex-minicart__item pa9 flex items-center justify-center relative bg-white">
-      <span className="f5">{label}</span>
+    <div className="vtex-minicart__item pa9 flex items-center justify-center relative bg-base">
+      <span className="t-body">{label}</span>
     </div>
   )
 
@@ -167,7 +167,7 @@ class MiniCartContent extends Component {
     const classes = classNames(
       'vtex-minicart__content overflow-x-hidden pa1',
       {
-        'vtex-minicart__content--small bg-white': !large,
+        'vtex-minicart__content--small bg-base': !large,
         'overflow-y-auto': large,
         'overflow-y-scroll': items.length > 3 && !large,
         'overflow-y-hidden': items.length <= 3 && !large,
@@ -203,7 +203,9 @@ class MiniCartContent extends Component {
           ))}
         </div>
         <div
-          className="vtex-minicart-content__footer w-100 bg-white pa4 bt b--silver pt4 flex flex-column items-end">
+          className="vtex-minicart-content__footer w-100 bg-base pa4 bt b--muted-3 pt4 flex flex-column items-end"
+        >
+
           {showDiscount && (
             <div className="vtex-minicart__content-discount blue w-100 flex justify-end items-center">
               <span className="ttl c-action-primary">{labelDiscount}</span>
@@ -237,7 +239,7 @@ class MiniCartContent extends Component {
   }
 
   renderLoading = () => (
-    <div className="vtex-minicart__item pa4 flex items-center justify-center relative bg-white">
+    <div className="vtex-minicart__item pa4 flex items-center justify-center relative bg-base">
       <Spinner />
     </div>
   )