+++ react/components/Sidebar.js
@@ -49,21 +49,22 @@ class Sidebar extends Component {
         <div style={{ willChange: 'opacity' }} className={scrimClasses} onClick={onOutsideClick} />
 
         <Animation
-          className="vtex-minicart__sidebar w-80 w-auto-ns h-100 fixed top-0 right-0 z-9999 bg-white shadow-2 flex flex-column"
+                    className="vtex-minicart__sidebar w-100 w-auto-ns h-100 fixed top-0 right-0 z-9999 bg-base shadow-2 flex flex-column"
+
           isActive={isOpen}
           type="drawerLeft"
         >
-          <div className="vtex-minicart__sidebar-header pointer flex flex-row items-center pa5 h3 bg-white w-100 z-max bb b--silver bw1">
+          <div className="vtex-minicart__sidebar-header pointer flex flex-row items-center pa5 h3 bg-base w-100 z-max bb b--muted-3 bw1">
             <div
-              className="mid-gray pa4 flex items-center"
+              className="c-muted-1 pa4 flex items-center"
               onClick={onOutsideClick}
             >
               <IconCaretRight size={17} />
             </div>
             <MiniCart
               hideContent
-              iconClasses="mid-gray"
-              labelClasses="mid-gray"
+              iconClasses="c-muted-1"
+              labelClasses="c-muted-1"
               iconLabel={intl.formatMessage({ id: 'sidebar-title' })}
             />
           </div>